Move model base plugin test to functional test
